Move cats.free.Free.Trampoline alias to cats.free package object
