+++ core/src/main/scala/cats/free/package.scala
@@ -1,0 +1,9 @@
+package cats
+
+package object free {
+
+  /** Alias for the free monad over the `Function0` functor. */
+  type Trampoline[A] = Free[Function0, A]
+
+  type FreeC[S[_], A] = Free[Coyoneda[S, ?], A]
+}
